pyproject.toml
[build-system]
requires = [
    "hatchling>=1.6.0",
]
build-backend = "hatchling.build"

[project]
name = "airflow-provider-grafana-loki"
description = "A provider package for pushing and reading airflow task log from Grafana Loki."
readme = "README.md"
license = "Apache License 2.0"
requires-python = "~=3.7"
authors = [
    { name = "Snjypl", email = "toepoe.py@gmail.com" },
]
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]
dependencies = [
    "apache-airflow>=2.0.0",
]
dynamic = [
    "version",
]

[project.optional-dependencies]
test = [
    "pytest-mock",
    "pytest>=7.0.0",
    "requests-mock",
]

[project.entry-points.apache_airflow_provider]
provider_info = "grafana_loki_provider.__init__:get_provider_info"

[project.urls]
Homepage = "http://grafana.com/"

[tool.hatch.version]
path = "grafana_loki_provider/__init__.py"
pattern = '.*?"versions":.*?\["(?P<version>.*?)"]'

[tool.hatch.build.targets.sdist]
include = [
    "/grafana_loki_provider",
]
